[MOD] Adapt <mrp_product_variants> extracting <mrp_product_variants_operations>
