+++ mrp_operations_extension/README.rst
@@ -1,0 +1,9 @@
+MRP operations extension module
+===============================
+
+This module adds:
+
+- New table to store operations to avoid typing them again.
+- Adds a relation from WorkcenterLines to BOM Lists.
+- Adds a relation from WorkcenterLines to Manufacturing Orders in Scheduled/Consumed/Finished Products.
+- Adds a relation between Routing Work Center Lines and Work Center extra Info.
