+++ mrp_operations_extension/views/mrp_bom_view.xml
@@ -2,13 +2,16 @@
 <openerp>
     <data>
 
-    <record id="mrp_bom_form_view_inh" model="ir.ui.view">
+        <record id="mrp_bom_form_view_inh" model="ir.ui.view">
             <field name="name">mrp.bom.form.inh</field>
             <field name="model">mrp.bom</field>
             <field name="inherit_id" ref="mrp.mrp_bom_form_view" />
             <field name="arch" type="xml">
-                <xpath expr="//tree/field[@name='product_id']" position="after">
-                   <field name="operation" domain="[('routing_id', '=', parent.routing_id)]"/>
+                <xpath expr="//tree/field[@name='product_id']"
+                    position="after">
+                    <field name="operation"
+                        domain="[('routing_id', '=', parent.routing_id)]"
+                        groups="mrp.group_mrp_routings" />
                 </xpath>
             </field>
         </record>
