+++ mrp_product_variants_operations/README.rst
@@ -1,0 +1,4 @@
+This is an auto install module.
+
+It is required for mrp_operations_extension and mrp_product_variants to work
+properly when both are installed.
